+++ src/plugins/backgrounds/apod/ApodTitle.tsx
@@ -4,7 +4,7 @@ import { Image } from "./types";
 type Props = Pick<Image, "title" | "copyright">;
 
 const Credit: React.FC<Props> = ({ title, copyright }) => (
-  <div className="title" style={{ lineHeight: 0 }}>
+  <div className="apod-credit title" style={{ lineHeight: 0 }}>
     <p>{title}</p>
     {copyright && <p style={{ textAlign: "right" }}>&copy; {copyright}</p>}
   </div>